refactor: Refactored ActivityGenerator's project structure
First working version of the app was all in one file: main.go
Now the project has been refactored and logic has been placed
in corresponding sub directories under app: controllers, services, models
