+++ app/models/five_minute_activity.go
@@ -1,0 +1,9 @@
+package models
+
+type FiveMinuteActivity struct {
+	standHrs        int
+	totalSteps      int
+	ExerciseMinutes int
+	caloriesBurned  float64
+	description     string
+}
